Split form definitions from selection coordination
Committed-By-Agent: codex
Committed-By-Agent: codex
Committed-By-Agent: codex
