+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/verticalmode/VerticalModeInitialScreenFactory.kt
@@ -1,18 +1,14 @@
 package com.stripe.android.paymentsheet.verticalmode
 
-import androidx.lifecycle.viewModelScope
 import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
 import com.stripe.android.lpmfoundations.paymentmethod.WalletType
 import com.stripe.android.paymentsheet.CustomerStateHolder
-import com.stripe.android.paymentsheet.DefaultFormHelper
+import com.stripe.android.paymentsheet.DefaultFormDefinitionFactory
 import com.stripe.android.paymentsheet.FormHelper
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen
 import com.stripe.android.paymentsheet.repositories.PaymentMethodMessagePromotionsHelper
-import com.stripe.android.paymentsheet.utils.childScope
 import com.stripe.android.paymentsheet.viewmodels.BaseSheetViewModel
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.cancel
 
 internal object VerticalModeInitialScreenFactory {
     fun create(
@@ -60,16 +56,10 @@ internal object VerticalModeInitialScreenFactory {
             (viewModel.selection.value as? PaymentSelection.New?)?.let { newPaymentSelection ->
                 val paymentMethodCode = newPaymentSelection.paymentMethodCreateParams.typeCode
 
-                // This form helper only answers formTypeForCode synchronously and is then discarded, so give it a
-                // scope we cancel immediately rather than leaking its init collector on viewModelScope for the life
-                // of the sheet.
-                val formTypeScope = viewModel.viewModelScope.childScope(Dispatchers.Main)
-                val formType = DefaultFormHelper.create(
+                val formType = DefaultFormDefinitionFactory.create(
                     viewModel = viewModel,
-                    coroutineScope = formTypeScope,
                     paymentMethodMetadata = paymentMethodMetadata
                 ).formTypeForCode(paymentMethodCode)
-                formTypeScope.cancel()
 
                 if (formType == FormHelper.FormType.UserInteractionRequired) {
                     add(
